minor changes
git-svn-id: file:///home/svn/shoola-omero/trunk@591 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/rnd/RenderingAgt.java
@@ -362,7 +362,7 @@ public class RenderingAgt
 	 * @param family		constant that identifies a transformation.
 	 * @param resolution	bit-resolution.
 	 */
-	void setQuantumStrategy(int k, int family, int resolution)
+	void setQuantumStrategy(double k, int family, int resolution)
 	{
 		renderingControl.setQuantumStrategy(family, k, resolution);
 		refreshImage();
